put to_pipeline_spec on ComponentSpec
