+++ sdk/python/kfp/compiler/helpers.py
@@ -12,9 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Helper functions for DSL compiler."""
+import collections
 import re
 from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
 
+import kfp
 from google.protobuf import json_format
 from kfp import dsl
 from kfp.compiler import pipeline_spec_builder as builder
@@ -366,3 +368,75 @@ def validate_pipeline_name(name: str) -> None:
             'Please specify a pipeline name that matches the regular '
             'expression "^[a-z0-9][a-z0-9-]{0,127}$" using '
             '`dsl.pipeline(name=...)` decorator.' % name)
+
+
+def create_pipeline_spec_for_component(
+        pipeline_name: str, pipeline_args: List[dsl.PipelineChannel],
+        task_group: tasks_group.TasksGroup) -> pipeline_spec_pb2.PipelineSpec:
+    """Creates a pipeline spec object for a component (single-component
+    pipeline).
+
+    Args:
+        pipeline_name: The pipeline name.
+        pipeline_args: The PipelineChannel arguments to the pipeline.
+        task_group: The pipeline's single task group (containing a single
+            task).
+
+    Returns:
+        A PipelineSpec proto representing the compiled pipeline.
+
+    Raises:
+        ValueError: If the argument is of unsupported types.
+    """
+
+    # this is method is essentially a simplified version
+    # of _create_pipeline_spec
+
+    # one-by-one building up the arguments for self._build_spec_by_group
+    validate_pipeline_name(pipeline_name)
+
+    pipeline_spec = pipeline_spec_pb2.PipelineSpec()
+    pipeline_spec.pipeline_info.name = pipeline_name
+    pipeline_spec.sdk_version = f'kfp-{kfp.__version__}'
+    # Schema version 2.1.0 is required for kfp-pipeline-spec>0.1.13
+    pipeline_spec.schema_version = '2.1.0'
+    pipeline_spec.root.CopyFrom(
+        builder.build_component_spec_for_group(
+            pipeline_channels=pipeline_args,
+            is_root_group=True,
+        ))
+
+    deployment_config = pipeline_spec_pb2.PipelineDeploymentConfig()
+    root_group = task_group
+
+    task_name_to_parent_groups, group_name_to_parent_groups = get_parent_groups(
+        root_group)
+
+    def get_inputs(task_group: tasks_group.TasksGroup,
+                   task_name_to_parent_groups):
+        inputs = collections.defaultdict(set)
+        if len(task_group.tasks) != 1:
+            raise ValueError(
+                f'Error compiling component. Expected one task in task group, got {len(task_group.tasks)}.'
+            )
+        only_task = task_group.tasks[0]
+        if only_task.channel_inputs:
+            for group_name in task_name_to_parent_groups[only_task.name]:
+                inputs[group_name].add((only_task.channel_inputs[-1], None))
+        return inputs
+
+    inputs = get_inputs(task_group, task_name_to_parent_groups)
+
+    build_spec_by_group(
+        pipeline_spec=pipeline_spec,
+        deployment_config=deployment_config,
+        group=root_group,
+        inputs=inputs,
+        dependencies={},  # no dependencies for single-component pipeline
+        rootgroup_name=root_group.name,
+        task_name_to_parent_groups=task_name_to_parent_groups,
+        group_name_to_parent_groups=group_name_to_parent_groups,
+        name_to_for_loop_group={},  # no for loop for single-component pipeline
+    )
+
+    return pipeline_spec
